+++ e2e/fixtures/compilers/new-babel/compiler.js
@@ -9,6 +9,7 @@ let logger;
 const compiler = {
   init: ({ rawConfig, dynamicConfig, api }) => {
     logger = api.getLogger();
+    return { write: true };
   },
   getDynamicConfig: ({ rawConfig }) => {
     const dynamicConfig = Object.assign({}, rawConfig);