+++ src/consumer/bit-map/component-map.js
@@ -1,10 +1,9 @@
 /** @flow */
 import R from 'ramda';
 import fs from 'fs-extra';
-import format from 'string-format';
 import path from 'path';
 import logger from '../../logger/logger';
-import { COMPONENT_ORIGINS, BIT_MAP, COMPONENT_DIR } from '../../constants';
+import { COMPONENT_ORIGINS, BIT_MAP } from '../../constants';
 import { pathNormalizeToLinux, pathJoinLinux, pathRelativeLinux, isValidPath } from '../../utils';
 import type { PathOsBasedRelative, PathLinux, PathOsBased } from '../../utils/path';
 import { Consumer } from '..';
@@ -14,6 +13,7 @@ import { NoFiles, EmptyDirectory } from '../component-ops/add-components/excepti
 import GeneralError from '../../error/general-error';
 import ValidationError from '../../error/validation-error';
 import ComponentNotFoundInPath from '../component/exceptions/component-not-found-in-path';
+import ConfigDir from './config-dir';
 
 export type ComponentOrigin = $Keys<typeof COMPONENT_ORIGINS>;
 
@@ -29,7 +29,7 @@ export type ComponentMapData = {
   mainFile: PathLinux,
   rootDir?: ?PathLinux,
   trackDir?: ?PathLinux,
-  configDir?: ?PathLinux,
+  configDir?: ?PathLinux | ?ConfigDir | string,
   origin: ComponentOrigin,
   detachedCompiler?: ?boolean,
   detachedTester?: ?boolean,
@@ -51,7 +51,7 @@ export default class ComponentMap {
   // be relative to consumer-root. (we can't save in the model relative to rootDir, otherwise the
   // dependencies paths won't work).
   trackDir: ?PathLinux; // relevant for AUTHORED only when a component was added as a directory, used for tracking changes in that dir
-  configDir: ?PathLinux;
+  configDir: ?ConfigDir;
   origin: ComponentOrigin;
   dependencies: ?(string[]); // needed for the link process
   mainDistFile: ?PathLinux; // needed when there is a build process involved
@@ -75,12 +75,18 @@ export default class ComponentMap {
     detachedCompiler,
     detachedTester
   }: ComponentMapData) {
+    let confDir;
+    if (configDir && typeof configDir === 'string') {
+      confDir = new ConfigDir(configDir);
+    } else if (configDir && configDir instanceof ConfigDir) {
+      confDir = configDir.clone();
+    }
     this.id = id;
     this.files = files;
     this.mainFile = mainFile;
     this.rootDir = rootDir;
     this.trackDir = trackDir;
-    this.configDir = configDir;
+    this.configDir = confDir;
     this.origin = origin;
     this.dependencies = dependencies;
     this.mainDistFile = mainDistFile;
@@ -91,6 +97,29 @@ export default class ComponentMap {
 
   static fromJson(componentMapObj: ComponentMapData): ComponentMap {
     return new ComponentMap(componentMapObj);
+  }
+
+  toPlainObject(): Object {
+    let res = {
+      id: this.id,
+      files: this.files,
+      mainFile: this.mainFile,
+      rootDir: this.rootDir,
+      trackDir: this.trackDir,
+      configDir: this.configDir ? this.configDir.linuxDirPath : undefined,
+      origin: this.origin,
+      dependencies: this.dependencies,
+      mainDistFile: this.mainDistFile,
+      originallySharedDir: this.originallySharedDir,
+      detachedCompiler: this.detachedCompiler,
+      detachedTester: this.detachedTester,
+      exported: this.exported
+    };
+    const notNil = (val) => {
+      return !R.isNil(val);
+    };
+    res = R.filter(notNil, res);
+    return res;
   }
 
   static getPathWithoutRootDir(rootDir: PathLinux, filePath: PathLinux): PathLinux {
@@ -239,7 +268,7 @@ export default class ComponentMap {
       return;
     }
     this.markBitMapChangedCb();
-    this.configDir = pathNormalizeToLinux(val);
+    this.configDir = new ConfigDir(val);
   }
 
   /**
@@ -247,8 +276,8 @@ export default class ComponentMap {
    */
   getBaseConfigDir(): ?PathLinux {
     if (!this.configDir) return null;
-    const trackDir = this.getTrackDir();
-    const configDir = format(this.configDir, { [COMPONENT_DIR]: trackDir, ENV_TYPE: '' });
+    const componentDir = this.getComponentDir();
+    const configDir = this.configDir && this.configDir.getResolved({ componentDir }).getEnvTypeCleaned().linuxDirPath;
     return configDir;
   }
 
@@ -302,6 +331,7 @@ export default class ComponentMap {
   }
 
   clone() {
+    // $FlowFixMe - there is some issue with the config dir type
     return new ComponentMap(this);
   }