+++ src/consumer/component-ops/inject-conf.js
@@ -7,6 +7,7 @@ import { removeEmptyDir, pathNormalizeToLinux } from '../../utils';
 import GeneralError from '../../error/general-error';
 import { COMPONENT_DIR } from '../../constants';
 import BitMap from '../bit-map';
+import ConfigDir from '../bit-map/config-dir';
 
 export type InjectConfResult = { id: string };
 
@@ -14,15 +15,14 @@ export default (async function injectConf(
   component: ConsumerComponent,
   consumerPath: string,
   bitMap: BitMap,
-  configDir: string,
+  configDir: ConfigDir,
   force?: boolean = false
 ): Promise<InjectConfResult> {
   const componentMap = component.componentMap;
-  const linuxConfigDir = pathNormalizeToLinux(configDir);
   if (!componentMap) {
     throw new GeneralError('could not find component in the .bitmap file');
   }
-  const trackDir = componentMap.getTrackDir();
+  const componentDir = componentMap.getComponentDir();
 
   // TODO: check if files were modified before deleting them and delete them only if forces provided
   if (!force) {
@@ -32,18 +32,17 @@ export default (async function injectConf(
   }
 
   // Passing here the ENV_TYPE as well to make sure it's not removed since we need it later
-  const resolvedConfigDir = format(linuxConfigDir, { [COMPONENT_DIR]: trackDir, ENV_TYPE: '{ENV_TYPE}' });
-  const resolvedConfigDirFullPath = path.normalize(path.join(consumerPath, resolvedConfigDir));
-
+  const resolvedConfigDir = configDir.getResolved({ componentDir });
   const deleteCompilerFilesP = component.compiler ? component.compiler.removeFilesFromFs() : Promise.resolve('');
   const deleteTesterFilesP = component.tester ? component.tester.removeFilesFromFs() : Promise.resolve('');
 
   await Promise.all([deleteCompilerFilesP, deleteTesterFilesP]);
 
   // Delete bit.json and bit.json dir
-  const bitJsonDir = format(resolvedConfigDirFullPath, { ENV_TYPE: '' });
-  await ComponentBitJson.removeIfExist(bitJsonDir);
-  await removeEmptyDir(bitJsonDir);
+  const bitJsonDir = resolvedConfigDir.getEnvTypeCleaned();
+  const bitJsonDirFullPath = path.normalize(path.join(consumerPath, bitJsonDir.dirPath));
+  await ComponentBitJson.removeIfExist(bitJsonDirFullPath);
+  await removeEmptyDir(bitJsonDirFullPath);
 
   return { id: component.id.toStringWithoutVersion() };
 });