+++ src/scope/exceptions/hash-not-found.js
@@ -1,4 +1,10 @@
 /** @flow */
 import AbstractError from '../../error/abstract-error';
 
-export default class HashNotFound extends AbstractError {}
+export default class HashNotFound extends AbstractError {
+  hash: string;
+  constructor(hash: string) {
+    super();
+    this.hash = hash;
+  }
+}