+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - support dynamic namespaces (replaced the box+name format with a dynamic name that can have multiple slashes to indicate a hierarchical namespace)
 - new eject-conf command to write bit.json and envs config files to file system
 - new inject-conf command to delete bit.json and envs config files from the file system
+- support return `{ write: true }` from environment init function
 
 ## [[13.0.4] - 2018-07-24](https://github.com/teambit/bit/releases/tag/v13.0.4)